+++ src/utils/cliError.ts
@@ -1,0 +1,13 @@
+export class CliError extends Error {
+    readonly exitCode: number;
+
+    constructor(message: string, exitCode = 1) {
+        super(message);
+        this.name = 'CliError';
+        this.exitCode = exitCode;
+    }
+}
+
+export const fail = (message: string, exitCode = 1): never => {
+    throw new CliError(message, exitCode);
+};